Fixing DnDFileUploader
*Refactored entire Uploader due to DropTarget not working from react-dnd library. Now utilizing the useDrop hook.
* Simplified the file uploader rendering logic.
* Dropzone is now DropZoneComposer (HOC) and no longer cares about what is rendered in it to provide more widespread use outside of file scenarios. Composer passes props from useDrop hook to composed component for consumption.
