+++ packages/components/drag-and-drop/index.js
@@ -1,5 +1,5 @@
 export { DndProvider, DropTarget, DragSource, DragLayer } from 'react-dnd';
 export { HTML5Backend, NativeTypes } from 'react-dnd-html5-backend';
-export { default as Dropzone } from './src/Dropzone';
-export { dropzoneTarget, dropzoneCollect } from './src/utils/dropzoneUtils';
+export { default as DropzoneComposer } from './src/DropzoneComposer';
+export { dropzoneSpec, dropzoneCollect } from './src/utils/dropzoneUtils';
 export { dragSourceCollect, dragSourceSpec } from './src/utils/dragSourceUtils';
