+++ packages/components/form-elements/src/styles/dnd-file-uploader.scss
@@ -13,14 +13,20 @@
 @import '~@epr0t0type/bankai-ui-buttons/src/styles/button.scss';
 
 .bankai-dnd-file-uploader {
+    min-height: 138px;
+
     &__inner {
         @include bankai-flex-column;
-        min-height: 138px;
         text-align: center;
         box-sizing: border-box;
         border-radius: var(--bankai-dnd-dropzone-border-radius, $bankai-form-input-border-radius);
         border: 2px dashed var(--bankai-neutral-border-color, $bankai-neutral-border-color);
         padding: 1.5rem;
+    }
+
+    .bankai-dropzone,
+    &__inner {
+        min-height: 138px;
     }
 
     &__content-container {
